instance-requirements.txt
Flask

# Flask Extensions
Flask-Admin
Flask-Assets
Flask-Cache
Flask-DebugToolbar
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-Security
Flask-Script
Flask-WTF

# Other
alembic
itsdangerous
cssmin
jsmin
python-dateutil
psycopg2

# Testing
pytest
pytest-cov
mccabe
flake8
